refactor: use constantly defined package version
